backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
groq==0.11.0
gspread==6.1.4
google-auth==2.35.0
pydantic==2.9.2
pydantic-settings==2.5.2
python-dotenv==1.0.1
twilio>=8.0.0
python-multipart>=0.0.9
